# Conflicts:
#	CODEOWNERS
#	buildconfig/appcompile/compile-sellerapp.gradle
#	buildconfig/dependencies/dependencies.gradle
#	features/merchant/seller_home/src/main/java/com/tokopedia/centralizedpromo/view/PromoCreationStaticData.kt
#	features/merchant/seller_home/src/main/res/values/strings.xml
#	libraries/applink/src/main/java/com/tokopedia/applink/DeeplinkMapper.kt
#	libraries/applink/src/main/java/com/tokopedia/applink/internal/ApplinkConstInternalSellerapp.kt
#	libraries/applink/src/main/res/values/internal_applink_sellerapp.xml

+++ features/travel/hotel/src/main/java/com/tokopedia/hotel/destination/view/fragment/HotelRecommendationFragment.kt
@@ -236,7 +236,7 @@ class HotelRecommendationFragment : BaseListFragment<PopularSearch, PopularSearc
     override fun onItemClicked(recentSearch: RecentSearch) {
         val intent = Intent()
         intent.putExtra(HOTEL_DESTINATION_NAME, recentSearch.property.value)
-        intent.putExtra(HOTEL_DESTINATION_ID, recentSearch.property.id.toInt())
+        intent.putExtra(HOTEL_DESTINATION_ID, recentSearch.property.id.toLong())
         intent.putExtra(HOTEL_DESTINATION_TYPE, recentSearch.property.type)
         activity?.setResult(Activity.RESULT_OK, intent)
         activity?.finish()
