+++ features/travel/hotel/src/main/java/com/tokopedia/hotel/homepage/presentation/activity/HotelHomepageActivity.kt
@@ -6,6 +6,7 @@ import android.os.Bundle
 import androidx.fragment.app.Fragment
 import com.tokopedia.abstraction.common.di.component.HasComponent
 import com.tokopedia.hotel.HotelComponentInstance
+import com.tokopedia.hotel.R
 import com.tokopedia.hotel.common.presentation.HotelBaseActivity
 import com.tokopedia.hotel.homepage.di.DaggerHotelHomepageComponent
 import com.tokopedia.hotel.homepage.di.HotelHomepageComponent
@@ -13,7 +14,7 @@ import com.tokopedia.hotel.homepage.presentation.fragment.HotelHomepageFragment
 
 class HotelHomepageActivity : HotelBaseActivity(), HasComponent<HotelHomepageComponent> {
 
-    private var id: Int = 0
+    private var id: Long = 0
     private var name: String = ""
     private var type: String = ""
     private var checkIn: String = ""
@@ -29,19 +30,19 @@ class HotelHomepageActivity : HotelBaseActivity(), HasComponent<HotelHomepageCom
         val uri = intent.data
         if (uri != null) {
             if (!uri.getQueryParameter(PARAM_HOTEL_ID).isNullOrEmpty()) {
-                id = uri.getQueryParameter(PARAM_HOTEL_ID).toInt()
+                id = uri.getQueryParameter(PARAM_HOTEL_ID).toLong()
                 name = uri.getQueryParameter(PARAM_HOTEL_NAME)
                 type = TYPE_PROPERTY
             } else if (!uri.getQueryParameter(PARAM_CITY_ID).isNullOrEmpty()) {
-                id = uri.getQueryParameter(PARAM_CITY_ID).toInt()
+                id = uri.getQueryParameter(PARAM_CITY_ID).toLong()
                 name = uri.getQueryParameter(PARAM_CITY_NAME)
                 type = TYPE_CITY
             } else if (!uri.getQueryParameter(PARAM_DISTRICT_ID).isNullOrEmpty()) {
-                id = uri.getQueryParameter(PARAM_DISTRICT_ID).toInt()
+                id = uri.getQueryParameter(PARAM_DISTRICT_ID).toLong()
                 name = uri.getQueryParameter(PARAM_DISTRICT_NAME)
                 type = TYPE_DISTRICT
             } else if (!uri.getQueryParameter(PARAM_REGION_ID).isNullOrEmpty()) {
-                id = uri.getQueryParameter(PARAM_REGION_ID).toInt()
+                id = uri.getQueryParameter(PARAM_REGION_ID).toLong()
                 name = uri.getQueryParameter(PARAM_REGION_NAME)
                 type = TYPE_REGION
             }
