+++ features/travel/hotel/src/main/java/com/tokopedia/hotel/hoteldetail/presentation/activity/HotelReviewActivity.kt
@@ -26,7 +26,7 @@ class HotelReviewActivity: HotelBaseActivity(), HasComponent<HotelDetailComponen
     }
 
     override fun getNewFragment(): Fragment =
-            HotelReviewFragment.createInstance(intent.getIntExtra(HotelReviewFragment.ARG_PROPERTY_ID, 0))
+            HotelReviewFragment.createInstance(intent.getLongExtra(HotelReviewFragment.ARG_PROPERTY_ID, 0))
 
     override fun getComponent(): HotelDetailComponent =
             DaggerHotelDetailComponent.builder()
@@ -36,7 +36,7 @@ class HotelReviewActivity: HotelBaseActivity(), HasComponent<HotelDetailComponen
     override fun getScreenName(): String = "Ulasan"
 
     companion object {
-        fun getCallingIntent(context: Context, propertyId: Int): Intent = Intent(context, HotelReviewActivity::class.java)
+        fun getCallingIntent(context: Context, propertyId: Long): Intent = Intent(context, HotelReviewActivity::class.java)
                 .putExtra(HotelReviewFragment.ARG_PROPERTY_ID,propertyId)
     }
